// src/generated/model/jsonBostotteSak.ts
/**
 * Generated by orval v6.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type {JsonBostotteSakKilde} from "./jsonBostotteSakKilde";
import type {JsonBostotteSakVedtaksstatus} from "./jsonBostotteSakVedtaksstatus";

export interface JsonBostotteSak {
    kilde?: JsonBostotteSakKilde;
    type?: string;
    dato?: string;
    status?: string;
    beskrivelse?: string;
    vedtaksstatus?: JsonBostotteSakVedtaksstatus;
}