// src/generated/model/jsonBostotteSakVedtaksstatus.ts
/**
 * Generated by orval v6.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type JsonBostotteSakVedtaksstatus =
    (typeof JsonBostotteSakVedtaksstatus)[keyof typeof JsonBostotteSakVedtaksstatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const JsonBostotteSakVedtaksstatus = {
    INNVILGET: "INNVILGET",
    AVSLAG: "AVSLAG",
    AVVIST: "AVVIST",
} as const;